requirements.txt
asgiref==3.4.0
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
boto3==1.17.92
certifi==2021.5.30
cffi==1.14.5
channels==2.1.5
chardet==4.0.0
click==8.0.1
constantly==15.1.0
cryptography==3.4.7
daphne==2.5.0
dj-database-url==0.5.0
Django==2.0
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
django-froala-editor==3.2.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-widget-tweaks==1.4.1
djangorestframework==3.7.3
fcm-django==0.3.10
gunicorn==20.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
jmespath==0.10.0
mysqlclient==2.0.3
Pillow==8.2.0
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfcm==1.4.9
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2017.3
requests==2.25.1
s3transfer==0.4.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.5
whitenoise==5.2.0
wrapt==1.12.1
zope.interface==5.4.0
